+++ app/src/main/java/madcourse/neu/edu/allot/participant/ParticipantActivity.java
@@ -1,7 +1,9 @@
-package madcourse.neu.edu.allot;
+package madcourse.neu.edu.allot.participant;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import madcourse.neu.edu.allot.R;
 
 public class ParticipantActivity extends AppCompatActivity {
