+++ app/src/main/java/madcourse/neu/edu/allot/task/TaskActivity.java
@@ -1,7 +1,9 @@
-package madcourse.neu.edu.allot;
+package madcourse.neu.edu.allot.task;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+
+import madcourse.neu.edu.allot.R;
 
 public class TaskActivity extends AppCompatActivity {
     @Override
