+++ app/src/main/res/layout/activity_place.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="madcourse.neu.edu.allot.PlaceActivity">
+    tools:context="madcourse.neu.edu.allot.place.PlaceActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appbar"
